302223 - Split "Stackato::Logyard" into Stackato::ApptailRegistrar and Stackato::LogHelper
Better reflect what the code is for. Also bump the version # as
the API changes here.
