+++ lib/cf/version.rb
@@ -1,3 +1,3 @@
 module Cf
-  VERSION = "3.0.0".freeze
+  VERSION = "3.0.1".freeze
 end
